H: Fixed issue validating registration
