+++ ats/src/main/resources/templates/index.html
@@ -20,7 +20,7 @@
 
         <!-- Logo Section -->
         <div class="container text-center" style="margin-top: 20px;">
-            <img th:src="@{/images/recruit-ats-logo.png}" alt="ATS Logo" style="max-width: 200px; margin: 20px 0;"/>
+            <img th:src="@{/images/ats.jpg}" alt="ATS Logo" style="max-width: 200px; margin: 20px 0;"/>
         </div>
 
         <!-- Footer -->
