+++ ats/src/main/resources/templates/layouts/common.html
@@ -8,7 +8,7 @@
                     <span class="icon-bar"></span>
                 </button>
                 <a class="navbar-brand" href="/" th:href="@{/}">
-                    <img src="/images/recruit-ats-logo.png" alt="ATS Logo" style="max-height: 30px;">
+                    <img src="/images/file.jpg" alt="ATS Logo" style="max-height: 30px;">
                 </a>
             </div>
             <div id="navbar" class="collapse navbar-collapse">
