+++ ats/src/main/resources/templates/register.html
@@ -1,50 +1,47 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraraq.net.nz/thymeleaf/layout"
-  layout:decorate="layouts/defaultTemplate">
-<body>
-
-<div layout:fragment="content">
-<form action="#" th:action="@{reLogin}" th:object="${loginModel}" method="post">
-			<table>
-				<tr>
-					<td><label>First Name:</label></td>
-                    <td><input type="text" th:field="*{fName}" ></td>
-                    <h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('fName')}" th:errors="*{fName}">First Name Error</h5>
-				</tr>
-				<tr>
-					<td><label>Last Name:</label></td>
-                    <td><input type="text" th:field="*{lName}" ></td>
-                    <h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('lName')}" th:errors="*{lName}">Last Name Error</h5>
-				</tr>
-				<tr>
-					<td><label>Username</label></td>
-					<td><input type="text" th:field="*{username}" ></td>
-					<h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('username')}" th:errors="*{username}">Username Error</h5>
-				</tr>
-				<tr>
-					<td><label>Password</label></td>
-					<td><input type="password" th:field="*{password}" ></td>
-					<h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('password')}" th:errors="*{password}">Password Error</h5>
-				</tr>
-				<tr>
-					<td><label>Email</label></td>
-					<td><input type="text" th:field="*{email}" ></td>
-					<h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('email')}" th:errors="*{email}">Email Error</h5>
-				</tr>
-				<tr>
-					<td><label>pNumber</label></td>
-					<td><input type="text" th:field="*{pNumber}" ></td>
-					<h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('pNumber')}" th:errors="*{pNumber}">Phone Number Error</h5>
-				</tr>
-				<tr>
-					<td style=text-align:center colspan="7"><input type="submit" value="Submit" /></td>
-				</tr>
-			</table>
-		<div th:if="${#fields.hasErrors('*')}">List Of Errors</div>
-		<div th:each="err : ${#fields.errors('*')}">
-			<span style="color:red" th:text="${err}" class="fieldError"></span>
+<html xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="layouts/defaultTemplate">
+	<body>
+		<div layout:fragment="content">
+			<form action="#" th:action="@{reLogin}" th:object="${userModel}" method="post">
+					<table>
+						<tr>
+							<td><label>First Name</label></td>
+							<td><input type="text" th:field="*{fName}" ></td>
+							<h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('fName')}" th:errors="*{fName}">First Name Error</h5>
+						</tr>
+						<tr>
+							<td><label>Last Name</label></td>
+							<td><input type="text" th:field="*{lName}" ></td>
+							<h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('lName')}" th:errors="*{lName}">Last Name Error</h5>
+						</tr>
+						<tr>
+							<td><label>Username</label></td>
+							<td><input type="text" th:field="*{username}" ></td>
+							<h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('username')}" th:errors="*{username}">Username Error</h5>
+						</tr>
+						<tr>
+							<td><label>Password</label></td>
+							<td><input type="password" th:field="*{password}" ></td>
+							<h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('password')}" th:errors="*{password}">Password Error</h5>
+						</tr>
+						<tr>
+							<td><label>Email</label></td>
+							<td><input type="text" th:field="*{email}" ></td>
+							<h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('email')}" th:errors="*{email}">Email Error</h5>
+						</tr>
+						<tr>
+							<td><label>Phone Number</label></td>
+							<td><input type="text" th:field="*{pNumber}" ></td>
+							<h5 class="alert alert-danger" style="color:red" th:if="${#fields.hasErrors('pNumber')}" th:errors="*{pNumber}">Phone Number Error</h5>
+						</tr>
+						<tr>
+							<td style=text-align:center colspan="7"><input type="submit" value="Submit" /></td>
+						</tr>
+					</table>
+				<div th:if="${#fields.hasErrors('*')}">List Of Errors</div>
+				<div th:each="err : ${#fields.errors('*')}">
+					<span style="color:red" th:text="${err}" class="fieldError"></span>
+				</div>
+			</form>
 		</div>
-</form>
-</div>
-</body>
+	</body>
 </html>
